+++ engine/servers/astronverse-executor/pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "astronverse-actionlib",
     "astronverse-websocket-server",
     "astronverse-tools",
+    "astronverse-workflowlib"
 ]
 
 [tool.uv.sources]
@@ -21,6 +22,7 @@ astronverse-baseline = {path = "../../shared/astronverse-baseline", editable = t
 astronverse-actionlib = {path = "../../shared/astronverse-actionlib", editable = true}
 astronverse-websocket-server = {path = "../../shared/astronverse-websocket-server", editable = true}
 astronverse-tools = {path = "../../shared/astronverse-tools", editable = true}
+astronverse-workflowlib = {path = "../../shared/astronverse-workflowlib", editable = true}
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/astronverse"]
