--- engine/servers/astronverse-executorv2/pyproject.toml
@@ -1,32 +1,0 @@
-[project]
-name = "astronverse-executor"
-version = "1.0.1"
-description = "astronverse-executor."
-requires-python = ">=3.13"
-
-dependencies = [
-    "websockets",
-    "psutil",
-    "anyio",
-    "requests",
-    "importlib_metadata",
-    "astronverse-baseline",
-    "astronverse-actionlib",
-    "astronverse-websocket-server",
-    "astronverse-tools",
-    "astronverse-workflowlib"
-]
-
-[tool.uv.sources]
-astronverse-baseline = {path = "../../shared/astronverse-baseline", editable = true}
-astronverse-actionlib = {path = "../../shared/astronverse-actionlib", editable = true}
-astronverse-websocket-server = {path = "../../shared/astronverse-websocket-server", editable = true}
-astronverse-tools = {path = "../../shared/astronverse-tools", editable = true}
-astronverse-workflowlib = {path = "../../shared/astronverse-workflowlib", editable = true}
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/astronverse"]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
